+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Implement mouse wheel input
 * RMB + wheel does advance and rewind frames (#594)
 * Add a Markers table in input editor
+* Hook AkAudio to enforce ALSA
 
 ### Changed